tests/conftest.py
import pytest
from httpx import AsyncClient
from httpx import ASGITransport
from app.main import app


@pytest.fixture
async def client():
    transport = ASGITransport(app=app)

    async with AsyncClient(
        transport=transport,
        base_url="http://test"
    ) as ac:
        yield ac